requirements.txt
streamlit
google-generativeai
pyttsx3
SpeechRecognition
deep-translator
pyaudio
